Move transformation swapping to traits.py
